requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.3.1
asyncprawcore==2.2.1
attrs==19.3.0
brotlipy==0.7.0
chardet==3.0.4
cryptography==3.4.7
discord-py==1.7.3
pip==21.1.3
praw==7.3.0
prawcore==2.2.0
pycodestyle==2.5.0
pycparser==2.20
pynacl==1.4.0
pyopenssl==20.0.1
pysocks==1.7.1
python-dotenv==0.18.0
typing-extensions==3.10.0.0
urllib3==1.25.9
websocket-client==0.57.0
yarl==1.4.2
youtube-dl==2021.6.6
